refactor(boilerplate-vite): dev-only servers with Vite HMR, remove prod build scripts
Both Express and Bun servers are dev tools with Vite middleware mode for
HMR on client and server. Production uses platform adapters (Vercel,
Cloudflare), not these servers.

- dev:express — Express + Vite SSR with renderToPipeableStream
- dev:bun — Bun.serve() + Vite SSR with renderToReadableStream
- dev — Vite client-only (no SSR)
- build:all — client build + storybook

Remove server build scripts, serve scripts, and separate dev server file.
