+++ apps/react/boilerplate-vite/package.json
@@ -6,17 +6,12 @@
   "license": "GPL-3.0-only",
   "scripts": {
     "dev": "vite",
-    "dev:ssr": "tsx src/server/server.express.dev.ts",
-    "build": "bun run build:client && bun run build:server",
-    "build:all": "bun run build:client && bun run build:server && bun run build:storybook",
+    "dev:express": "tsx src/server/server.express.ts",
+    "dev:bun": "bun src/server/server.bun.ts",
+    "build": "bun run build:client",
+    "build:all": "bun run build:client && bun run build:storybook",
     "build:storybook": "storybook build",
     "build:client": "vite build --ssrManifest --outDir dist/client",
-    "build:server": "bun run build:server:express",
-    "build:server:express": "vite build --ssr src/server/server.express.ts --outDir dist/server",
-    "build:server:bun": "vite build --ssr src/server/server.bun.ts --outDir dist/server",
-    "serve": "bun run serve:express",
-    "serve:express": "bun run build:client && bun run build:server:express && node dist/server/server.express.js",
-    "serve:bun": "bun run build:client && bun run build:server:bun && bun dist/server/server.bun.js",
     "check": "bun run check:biome && bun run check:ts",
     "check:fix": "bun run check:biome:fix && bun run check:ts",
     "check:biome": "biome check",
