+++ apps/react/boilerplate-vite/src/server/server.express.ts
@@ -1,64 +1,64 @@
-import fs from "node:fs/promises";
-import path from "node:path";
+/**
+ * Express development server with SSR and HMR.
+ *
+ * Vite handles client HMR and module transforms. Server modules are loaded
+ * via vite.ssrLoadModule() — changes to routes, pages, or server code are
+ * picked up without restarting the process.
+ *
+ * Production deployments use platform adapters (Vercel, Cloudflare, etc.),
+ * not this server.
+ */
+import fs from "node:fs";
 import * as process from "node:process";
-import { JSXRenderer, SitemapRenderer } from "@canonical/react-ssr/renderer";
-import { serveStream } from "@canonical/react-ssr/server";
 import express from "express";
-import { getAuthRedirectHref } from "../routes.js";
-import EntryServer, { type InitialData } from "./entry.js";
-import getSitemapItems from "./sitemap.js";
+import { createServer as createViteServer } from "vite";
 
-const PORT = process.env.PORT || 5173;
+const PORT = Number(process.env.PORT) || 5173;
 
-const htmlString = await fs.readFile(
-  path.join(process.cwd(), "dist", "client", "index.html"),
-  "utf-8",
-);
+async function start() {
+  const app = express();
 
-const app = express();
-
-app.use(/^\/(assets|public)/, express.static("dist/client/assets"));
-
-app.get("/sitemap.xml", async (_req, res) => {
-  const renderer = new SitemapRenderer([getSitemapItems], {
-    baseUrl: `http://localhost:${PORT}`,
-    defaultChangefreq: "monthly",
+  const vite = await createViteServer({
+    server: { middlewareMode: true },
+    appType: "custom",
   });
-  const { pipe } = renderer.renderToPipeableStream();
 
-  await renderer.statusReady;
-  res.setHeader("content-type", "application/xml; charset=utf-8");
-  res.status(renderer.statusCode);
-  pipe(res);
-});
+  app.use(vite.middlewares);
 
-app.use((req, res, next) => {
-  const requestUrl = req.url || "/";
-  const authRedirect = getAuthRedirectHref(requestUrl);
+  app.use(async (req, res, next) => {
+    const url = req.originalUrl || "/";
 
-  if (authRedirect) {
-    res.redirect(302, authRedirect);
+    try {
+      const template = fs.readFileSync("index.html", "utf-8");
+      const html = await vite.transformIndexHtml(url, template);
 
-    return;
-  }
+      const { default: EntryServer } = await vite.ssrLoadModule(
+        "/src/server/entry.tsx",
+      );
+      const { JSXRenderer } = await vite.ssrLoadModule(
+        "@canonical/react-ssr/renderer",
+      );
 
-  next();
-});
+      const renderer = new JSXRenderer(
+        EntryServer,
+        { url },
+        { htmlString: html },
+      );
+      const result = renderer.renderToPipeableStream();
 
-app.use(
-  serveStream((req) => {
-    const requestUrl = req.url || "/";
+      await renderer.statusReady;
+      res.status(renderer.statusCode);
+      res.setHeader("content-type", "text/html; charset=utf-8");
+      result.pipe(res);
+    } catch (error) {
+      vite.ssrFixStacktrace(error as Error);
+      next(error);
+    }
+  });
 
-    return new JSXRenderer(
-      EntryServer,
-      { url: requestUrl } satisfies InitialData,
-      { htmlString },
-    );
-  }),
-);
+  app.listen(PORT, () => {
+    console.log(`Express dev server on http://localhost:${PORT}/`);
+  });
+}
 
-app.listen(PORT, () => {
-  console.log(`Express server started on http://localhost:${PORT}/`);
-});
-
-export default app;
+start();
